+++ CSharp_Assignments.sln
@@ -11,7 +11,15 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "unit", "unit", "{C778E127-C
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "acceptance", "acceptance", "{93EBAE55-A8BA-40FF-93AA-7D9E4359B6B9}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Assignment7", "src\Assign7\Assignment7\Assignment7.csproj", "{1D153560-8D04-4B71-A196-4694EAAFF312}"
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Assign7", "Assign7", "{27F42C44-B214-4BDB-8014-BE44376E032C}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Assignment7", "src\Assign7\Assignment7\Assignment7.csproj", "{4C0C1A50-3235-4015-9560-FB3D4B9B4384}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "IDisposableDemo", "src\Assign7\IDisposableDemo\IDisposableDemo.csproj", "{AE875A52-D112-4356-9A6E-D8CCFAEBB700}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GarbageCollection", "src\Assign7\GarbageCollection\GarbageCollection.csproj", "{FE8F9DF7-5828-4293-A56C-9AB2E8A0F1E5}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ValueAndReferenceTypes", "src\Assign7\ValueAndReferenceTypes\ValueAndReferenceTypes.csproj", "{B865F787-DBC8-4749-801A-05D017526ABD}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -19,10 +27,22 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{1D153560-8D04-4B71-A196-4694EAAFF312}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{1D153560-8D04-4B71-A196-4694EAAFF312}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{1D153560-8D04-4B71-A196-4694EAAFF312}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{1D153560-8D04-4B71-A196-4694EAAFF312}.Release|Any CPU.Build.0 = Release|Any CPU
+		{4C0C1A50-3235-4015-9560-FB3D4B9B4384}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{4C0C1A50-3235-4015-9560-FB3D4B9B4384}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{4C0C1A50-3235-4015-9560-FB3D4B9B4384}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{4C0C1A50-3235-4015-9560-FB3D4B9B4384}.Release|Any CPU.Build.0 = Release|Any CPU
+		{AE875A52-D112-4356-9A6E-D8CCFAEBB700}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{AE875A52-D112-4356-9A6E-D8CCFAEBB700}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{AE875A52-D112-4356-9A6E-D8CCFAEBB700}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{AE875A52-D112-4356-9A6E-D8CCFAEBB700}.Release|Any CPU.Build.0 = Release|Any CPU
+		{FE8F9DF7-5828-4293-A56C-9AB2E8A0F1E5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{FE8F9DF7-5828-4293-A56C-9AB2E8A0F1E5}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{FE8F9DF7-5828-4293-A56C-9AB2E8A0F1E5}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{FE8F9DF7-5828-4293-A56C-9AB2E8A0F1E5}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B865F787-DBC8-4749-801A-05D017526ABD}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B865F787-DBC8-4749-801A-05D017526ABD}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B865F787-DBC8-4749-801A-05D017526ABD}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B865F787-DBC8-4749-801A-05D017526ABD}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -30,7 +50,11 @@ Global
 	GlobalSection(NestedProjects) = preSolution
 		{C778E127-C02C-4A37-8E85-3D86DCC063DC} = {06861454-9561-492A-9F93-A33B6F7AE3AB}
 		{93EBAE55-A8BA-40FF-93AA-7D9E4359B6B9} = {06861454-9561-492A-9F93-A33B6F7AE3AB}
-		{1D153560-8D04-4B71-A196-4694EAAFF312} = {DCAA095E-2796-41FA-A1D0-1134E29309E7}
+		{27F42C44-B214-4BDB-8014-BE44376E032C} = {DCAA095E-2796-41FA-A1D0-1134E29309E7}
+		{4C0C1A50-3235-4015-9560-FB3D4B9B4384} = {27F42C44-B214-4BDB-8014-BE44376E032C}
+		{AE875A52-D112-4356-9A6E-D8CCFAEBB700} = {27F42C44-B214-4BDB-8014-BE44376E032C}
+		{FE8F9DF7-5828-4293-A56C-9AB2E8A0F1E5} = {27F42C44-B214-4BDB-8014-BE44376E032C}
+		{B865F787-DBC8-4749-801A-05D017526ABD} = {27F42C44-B214-4BDB-8014-BE44376E032C}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {D4CECB66-8E11-49D9-AA1E-464EE0231450}
